requirements.txt
# Repositorio extra para soporte GPU (RTX 50 serie)
--extra-index-url https://download.pytorch.org/whl/cu128

# Librerías de IA (con soporte CUDA)
torch
torchvision
torchaudio

# Utilidades y Modelos
ultralytics
ddgs
requests
Pillow
opencv-python
numpy
